Change version from 0.4 to 0.5
git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk/sphinx3@3881 c60273ff-4a12-0410-8d58-82ceea6d4170

+++ src/libs3decoder/liblm/lm.c
@@ -1344,9 +1344,9 @@ int32 lm_rawscore (lm_t *lm, int32 score, float64 lwf)
 {
 
     if (lwf != 1.0)
-        score /= lwf;
+        score /= (int32)lwf;
     score -= lm->wip;
-    score /= lm->lw;
+    score /= (int32)lm->lw;
     
     return score;
 }
